backend/requirements.txt
asgiref==3.7.2
Django==5.0.3
graphql-core==3.2.3
phonenumbers==8.13.32
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
strawberry-graphql==0.220.0
strawberry-graphql-django==0.33.0
strawberry_jwt_auth==0.2.3
typing_extensions==4.10.0
